feat: add splitChunks custom config
